requirements.txt
einops==0.8.1
editdistance==0.8.1
# for pytorch-lightning CLI
jsonargparse[signatures]==4.40.2
# dev-dependency
flake8==7.3.0
black==25.1.0
isort==6.0.1
jupyter==1.1.1
opencv-python==4.12.0.88
matplotlib==3.10.5
# for test in crohme
typer==0.16.0
beautifulsoup4==4.13.4
lxml>=6.0.0
torch>=2.8.0
torchvision>=0.23.0
torchmetrics>=1.8.1
lightning>=2.5.2
pillow>=11.3.0
tensorboard>=2.20.0